Update qs package to address security advisory
- Updates the qs package to `^6.5.1`
- Address security advisory: https://snyk.io/vuln/npm:qs:20170213

This was raised when scanning fluxthis for vulnerabilities: https://snyk.io/test/npm/fluxthis?severity=high&severity=medium&severity=low

All tests continue to pass after updating this package. It's only used in one place and the API of the `qs` library hasn't changed for what Fluxthis uses it for.